Code cleanup & reorg.
Files checked in: DataRepo/tasks.py DataRepo/templates/DataRepo/search/downloads/download_progress.html DataRepo/templatetags/customtags.py DataRepo/urls.py DataRepo/views.py DataRepo/advanced_search_utils.py
[skip ci]
