+++ DataRepo/templates/DataRepo/search/downloads/download_progress.html
@@ -31,7 +31,7 @@
                                 setTimeout(updateProgress, 500, progressUrl, cnt);
                             } else {
                                 setProgress("Success. " + details.current + "/" + details.total + ".  Downloading...");
-                                window.location.replace(downloadUrl)
+                                window.location.replace(downloadUrl);
                             }
                         }
                     });
